20190719ソース

+++ aoyama-macchinetta/aoyama-macchinetta-web/src/main/java/co/jp/aoyama/macchinetta/app/delivery/ShopDeliveryController.java
@@ -19,8 +19,10 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import co.jp.aoyama.macchinetta.app.session.SessionContent;
+import co.jp.aoyama.macchinetta.domain.model.GeneralCode;
 import co.jp.aoyama.macchinetta.domain.model.ShopDelivery;
 import co.jp.aoyama.macchinetta.domain.service.delivery.ShopDeliveryService;
+import co.jp.aoyama.macchinetta.domain.service.generalcode.GeneralCodeService;
 
 /**
  * 納期管理
@@ -38,6 +40,9 @@ public class ShopDeliveryController {
     ShopDeliveryService shopDeliveryService;
     
     @Inject
+    GeneralCodeService generalCodeService;
+    
+    @Inject
     Mapper beanMapper;
 	
 	/**
@@ -46,6 +51,8 @@ public class ShopDeliveryController {
 	 */
 	@RequestMapping(value = "init", method = RequestMethod.GET)
 	public String pieceInit(Model model) {
+		List<GeneralCode> generalCodeList = getGeneralCode();
+		model.addAttribute("generalCodeList", generalCodeList);
 		return "delivery/shopDeliveryForm";
 	}
 	
@@ -65,6 +72,19 @@ public class ShopDeliveryController {
 		List<ShopDelivery> shopDeliveryDeleteList = new ArrayList<ShopDelivery>();
 		//戻り先
 		List<ShopDeliveryForm> shopDeliveryFormListReturn = new ArrayList<ShopDeliveryForm>();
+		//本方法では、データが間違っていると記録されている表示
+		boolean failureTemp = false;
+		
+		//店着納期区分のcodeValueを取得します
+		List<GeneralCode> generalCodeList = getGeneralCode();
+		for (int i = 0; i < shopDeliveryFormList.size(); i++) {
+			for (int j = 0; j < generalCodeList.size(); j++) {
+				if(generalCodeList.get(j).getCodeName().equals(shopDeliveryFormList.get(i).getShopDeliveryClass())) {
+					shopDeliveryFormList.get(i).setShopDeliveryClass(generalCodeList.get(j).getCodeValue());
+					break;
+				}
+			}
+		}
 		
 		for (int i = 0; i < shopDeliveryFormList.size(); i++) {
 			ShopDeliveryForm shopDeliveryFormGet = shopDeliveryFormList.get(i);
@@ -73,7 +93,26 @@ public class ShopDeliveryController {
 			if(shopDeliveryFormGet.getDelType() == true) {
 				//削除の操作
 				ShopDelivery shopDelivery = beanMapper.map(shopDeliveryFormGet, ShopDelivery.class);
-				shopDeliveryDeleteList.add(shopDelivery);
+				ShopDelivery shopDeliveryQueryByPrimaryKey = shopDeliveryService.shopDeliveryQueryByPrimaryKey(shopDelivery);
+				if(shopDeliveryQueryByPrimaryKey != null) {
+					if(shopDelivery.getVersion().equals(shopDeliveryQueryByPrimaryKey.getVersion())) {
+						//バージョン番号が正しいです、削除の作業
+						shopDeliveryDeleteList.add(shopDelivery);
+						shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+					}else {
+						//バージョン番号が不正です
+						//setUpdateFailure("-4")：バージョン番号が不正です
+						shopDeliveryFormGet.setUpdateFailure("-4");
+						failureTemp = true;
+						shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+					}
+				}else {
+					//納期情報の主キーはすでに存在しない。確認して再入力してください。
+		            //setUpdateFailure("-3")：削除のエラーが存在する場合
+					shopDeliveryFormGet.setUpdateFailure("-3");
+					failureTemp = true;
+		            shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+				}
 			}else{
 				//新規または修正の操作
 				ShopDelivery shopDelivery = beanMapper.map(shopDeliveryFormGet, ShopDelivery.class);
@@ -82,17 +121,27 @@ public class ShopDeliveryController {
 					//納期情報の有無の照会
 					ShopDelivery shopDeliveryQueryByPrimaryKey = shopDeliveryService.shopDeliveryQueryByPrimaryKey(shopDelivery);
 					if(shopDeliveryQueryByPrimaryKey != null) {
-						//納期情報が存在し、修正の作業
-						shopDelivery.setUpdatedAt(new Date());
-						shopDelivery.setUpdatedUserId(sessionContent.getUserId());
-						shopDeliveryUpdateList.add(shopDelivery);
-						//setUpdateFailure("0")：エラーが存在しない場合
-						shopDeliveryFormGet.setUpdateFailure("0");
-						shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+						//納期情報が存在し
+						if(shopDelivery.getVersion().equals(shopDeliveryQueryByPrimaryKey.getVersion())) {
+							//バージョン番号が正しいです、修正の作業
+							shopDelivery.setUpdatedAt(new Date());
+							shopDelivery.setUpdatedUserId(sessionContent.getUserId());
+							shopDeliveryUpdateList.add(shopDelivery);
+							//setUpdateFailure("0")：エラーが存在しない場合
+							shopDeliveryFormGet.setUpdateFailure("0");
+							shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+						}else {
+							//バージョン番号が不正です
+							//setUpdateFailure("-4")：バージョン番号が不正です
+							shopDeliveryFormGet.setUpdateFailure("-4");
+							failureTemp = true;
+				            shopDeliveryFormListReturn.add(shopDeliveryFormGet);
+						}
 					}else {
 						//納期情報の主キーはすでに存在しない。確認して再入力してください。
-			            //setUpdateFailure("-1")：エラーが存在する場合
-						shopDeliveryFormGet.setUpdateFailure("-1");
+			            //setUpdateFailure("-2")：修正のエラーが存在する場合
+						shopDeliveryFormGet.setUpdateFailure("-2");
+						failureTemp = true;
 			            shopDeliveryFormListReturn.add(shopDeliveryFormGet);
 					}
 				}else if("2".equals(shopDelivery.getOptionType())) {
@@ -101,8 +150,9 @@ public class ShopDeliveryController {
 					Integer shopDeliveryQueryIntersection = shopDeliveryService.shopDeliveryQueryIntersection(shopDelivery);
 					if(shopDeliveryQueryIntersection != 0) {
 						//納期情報の主キーはすでに存在している。確認して再入力してください。
-						//setUpdateFailure("-1")：エラーが存在する場合
+						//setUpdateFailure("-1")：新規のエラーが存在する場合
 						shopDeliveryFormGet.setUpdateFailure("-1");
+						failureTemp = true;
 						shopDeliveryFormListReturn.add(shopDeliveryFormGet);
 					}else {
 						//条件を満たす
@@ -118,23 +168,68 @@ public class ShopDeliveryController {
 				}
 			}
 		}
+		
+		//すべてのデータに誤りがないので、新規、修正、削除の操作を開始します
+		if(failureTemp == false) {
+			//画面に表示されているデータのバージョン番号の変更を返します
+			for (int i = 0; i < shopDeliveryFormListReturn.size(); i++) {
+				if(shopDeliveryFormListReturn.get(i).getVersion() == null) {
+					Short s1 = 1;
+					shopDeliveryFormListReturn.get(i).setVersion(s1);
+				}else {
+					Short version = shopDeliveryFormListReturn.get(i).getVersion();
+					shopDeliveryFormListReturn.get(i).setVersion(++version);
+				}
+			}
+			
+			if(shopDeliveryInsertList.size() != 0) {
+				//新規の数はではありません
+				//新規された場合、バージョン番号は１に設定されます
+				for (int i = 0; i < shopDeliveryInsertList.size(); i++) {
+					Short s1 = 1;
+					shopDeliveryInsertList.get(i).setVersion(s1);
+				}
+				
+				shopDeliveryService.insertShopDeliveryByPrimaryKey(shopDeliveryInsertList);
+			}
+			
+			if(shopDeliveryUpdateList.size() != 0) {
+				//修正の数はではありません
+				//修正の場合、バージョン番号は１から増加します
+				for (int i = 0; i < shopDeliveryUpdateList.size(); i++) {
+					Short version = shopDeliveryUpdateList.get(i).getVersion();
+					shopDeliveryUpdateList.get(i).setVersion(++version);
+				}
+				
+				shopDeliveryService.updateShopDeliveryByPrimaryKey(shopDeliveryUpdateList);
+			}
+			
+			if(shopDeliveryDeleteList.size() != 0) {
+				//削除の数はではありません
+				//戻ったりリストから消去したデータを削除します
+				for (int i = 0, length = shopDeliveryFormListReturn.size(); i < length; i++) {
+					if (shopDeliveryFormListReturn.get(i).getDelType() == true) {
+						shopDeliveryFormListReturn.remove(shopDeliveryFormListReturn.get(i));
+						length--;
+						i--;
+					}
+				}
+				
+				shopDeliveryService.deleteShopDeliveryByPrimaryKey(shopDeliveryDeleteList);
+			}
+		}
+		
+		//店着納期区分の名前を取得します
+		for (int i = 0; i < shopDeliveryFormListReturn.size(); i++) {
+			for (int j = 0; j < generalCodeList.size(); j++) {
+				if(generalCodeList.get(j).getCodeValue().equals(shopDeliveryFormListReturn.get(i).getShopDeliveryClass())) {
+					shopDeliveryFormListReturn.get(i).setCodeName(generalCodeList.get(j).getCodeName());
+					break;
+				}
+			}
+		}
 	
-	if(shopDeliveryInsertList.size() != 0) {
-		//新規の数はではありません
-		shopDeliveryService.insertShopDeliveryByPrimaryKey(shopDeliveryInsertList);
-	}
-	
-	if(shopDeliveryUpdateList.size() != 0) {
-		//修正の数はではありません
-		shopDeliveryService.updateShopDeliveryByPrimaryKey(shopDeliveryUpdateList);
-	}
-	
-	if(shopDeliveryDeleteList.size() != 0) {
-		//削除の数はではありません
-		shopDeliveryService.deleteShopDeliveryByPrimaryKey(shopDeliveryDeleteList);
-	}
-	
-	return shopDeliveryFormListReturn;
+		return shopDeliveryFormListReturn;
 	}
 		
 	/**
@@ -149,15 +244,17 @@ public class ShopDeliveryController {
 		List<ShopDeliveryForm> shopDeliveryFormListReturn = new ArrayList<ShopDeliveryForm>();
 		//日付書式の変更:string-->date
 		shopDeliveryForm = stringToDate(shopDeliveryForm);
+		//店着納期区分の名前を取得します
+		List<GeneralCode> generalCodeList = getGeneralCode();
 		
 		if("全区分".equals(shopDeliveryForm.getShopDeliveryClass())) {
 			//店着納期区分は全区分です
-			for (int i = 1; i <= 9; i++) { 
+			for (int i = 0; i < generalCodeList.size(); i++) { 
 				//ループ、ShopDeliveryFormの値を設定します
 				ShopDeliveryForm shopDeliveryFormFor = new ShopDeliveryForm();
 				shopDeliveryFormFor.setOrderOnStartDate(shopDeliveryForm.getOrderOnStartDate());
 				shopDeliveryFormFor.setOrderOnEndDate(shopDeliveryForm.getOrderOnEndDate());
-				shopDeliveryFormFor.setShopDeliveryClass("0" + String.valueOf(i));
+				shopDeliveryFormFor.setShopDeliveryClass(generalCodeList.get(i).getCodeValue());
 				shopDeliveryFormFor.setOrderOnStartDateStr(shopDeliveryForm.getOrderOnStartDateStr());
 				shopDeliveryFormFor.setOrderOnEndDateStr(shopDeliveryForm.getOrderOnEndDateStr());
 				shopDeliveryFormFor.setErrorIdentification(shopDeliveryForm.getErrorIdentification());
@@ -184,11 +281,11 @@ public class ShopDeliveryController {
 					shopDeliveryFormFor = dateToString(shopDeliveryFormFor);
 					
 					//setErrorIdentification("0")：エラーが存在しない場合
-					shopDeliveryForm.setErrorIdentification("0");
+					shopDeliveryFormFor.setErrorIdentification("0");
 					shopDeliveryFormListReturn.add(shopDeliveryFormFor); 
 				}else {
 					//setErrorIdentification("-1")：エラーが存在する場合
-					shopDeliveryForm.setErrorIdentification("-1");
+					shopDeliveryFormFor.setErrorIdentification("-1");
 					shopDeliveryFormListReturn.add(shopDeliveryFormFor); 
 				} 
 				
@@ -221,6 +318,16 @@ public class ShopDeliveryController {
 			} 
 			
 		}
+		
+		//店着納期区分の名前を取得します
+		for (int i = 0; i < shopDeliveryFormListReturn.size(); i++) {
+			for (int j = 0; j < generalCodeList.size(); j++) {
+				if(shopDeliveryFormListReturn.get(i).getShopDeliveryClass().equals(generalCodeList.get(j).getCodeValue())) {
+					shopDeliveryFormListReturn.get(i).setCodeName(generalCodeList.get(j).getCodeName());
+					break;
+				}
+			}
+		}
 		 
 		return shopDeliveryFormListReturn;
 	}
@@ -247,6 +354,18 @@ public class ShopDeliveryController {
 			shopDeliveryFormFor = dateToString(shopDeliveryFormFor);
 			shopDeliveryFormList.add(shopDeliveryFormFor);
 		}
+		
+		//店着納期区分の名前を取得します
+		List<GeneralCode> generalCodeList = getGeneralCode();
+		for (int i = 0; i < shopDeliveryFormList.size(); i++) {
+			for (int j = 0; j < generalCodeList.size(); j++) {
+				if(shopDeliveryFormList.get(i).getShopDeliveryClass().equals(generalCodeList.get(j).getCodeValue())) {
+					shopDeliveryFormList.get(i).setCodeName(generalCodeList.get(j).getCodeName());
+					break;
+				}
+			}
+		}
+		
 		return shopDeliveryFormList;
 	}
 	
@@ -383,6 +502,17 @@ public class ShopDeliveryController {
 			//条件の該当するデータがありません
 			shopDeliveryForm.setBlankIntervalEnd(null);
 		}
+	}
+	
+	/**
+	 * GeneralCodeのオブジェクトを取得
+	 * @return
+	 */
+	private List<GeneralCode> getGeneralCode(){
+		GeneralCode generalCode = new GeneralCode();
+		generalCode.setCodeType("01");
+		List<GeneralCode> generalCodeList = generalCodeService.selectGeneralCodeByCodeType(generalCode);
+		return generalCodeList;
 	}
 
 }
